1. move repository-path of room and booking 2. add DI of booking-memory-repo to booking-usecase
