+++ src/booking/domain/model/booking/inversify.config.ts
@@ -1,0 +1,8 @@
+import { Container } from "inversify";
+import { repositoryTypes } from "./types";
+import { BookingRepository } from "./repository";
+import { MemoryBookingRepository } from '../../../adapter/repository/booking/memory-repository';
+
+const container = new Container();
+container.bind<BookingRepository>(repositoryTypes.BookingRepository).to(MemoryBookingRepository);
+export { container };
