+++ src/booking/domain/model/booking/types.ts
@@ -1,0 +1,4 @@
+const repositoryTypes = {
+  BookingRepository: Symbol.for("BookingRepository")
+};
+export { repositoryTypes };
