+++ src/booking/usecase/booking/__tests__/checkin.spec.ts
@@ -19,14 +19,18 @@ import {
   CreateRoomOutput
 } from '../../room/index';
 import { PaymentMethod, BookingStatus } from '../../../domain/model/booking/booking';
-import { MemoryBookingRepository } from '../../../adapter/repository/booking/memory-repository';
+//import { MemoryBookingRepository } from '../../../adapter/repository/booking/memory-repository';
+import { BookingRepository } from '../../../domain/model/booking/repository';
+import { repositoryTypes } from '../../../domain/model/booking/types';
+import { container } from '../../../domain/model/booking/inversify.config';
+
 import { MemoryRoomRepository } from '../../../adapter/repository/room/memory-repository';
 
 describe('Checkin for booking', function () {
   let booking_id: string;
   let room_id: string;
   const roomRepo = new MemoryRoomRepository();
-  const bookingRepo = new MemoryBookingRepository();
+  const bookingRepo = container.get<BookingRepository>(repositoryTypes.BookingRepository);//new MemoryBookingRepository();
   // create a room first
   beforeEach(async () => {
     const roomNo = '103A';
