+++ src/booking/usecase/booking/__tests__/pay.spec.ts
@@ -16,14 +16,18 @@ import {
   CreateRoomOutput
 } from '../../room/index';
 import { PaymentMethod, BookingStatus } from '../../../domain/model/booking/booking';
-import { MemoryBookingRepository } from '../../../adapter/repository/booking/memory-repository';
+//import { MemoryBookingRepository } from '../../../adapter/repository/booking/memory-repository';
+import { BookingRepository } from '../../../domain/model/booking/repository';
+import { repositoryTypes } from '../../../domain/model/booking/types';
+import { container } from '../../../domain/model/booking/inversify.config';
+
 import { MemoryRoomRepository } from '../../../adapter/repository/room/memory-repository';
 
 describe('Pay for booking', function () {
   let booking_id: string;
   let room_id: string;
   const roomRepo = new MemoryRoomRepository();
-  const bookingRepo = new MemoryBookingRepository();
+  const bookingRepo = container.get<BookingRepository>(repositoryTypes.BookingRepository);//new MemoryBookingRepository();
   // create a room first
   beforeEach(async () => {
     const roomNo = '103A';
