+++ src/booking/usecase/booking/close.ts
@@ -1,4 +1,4 @@
-import { BookingRepository } from './repository';
+import { BookingRepository } from '../../domain/model/booking/repository';
 import { BookingId, Booking } from '../../domain/model/booking/booking';
 
 export class CloseBookingUsecase {
