+++ src/booking/usecase/booking/index.ts
@@ -1,5 +1,4 @@
 export * from './create';
-export * from './repository';
 export * from './get';
 export * from './edit';
 export * from './pay';
