--- src/booking/usecase/booking/repository.ts
@@ -1,7 +1,0 @@
-import { Booking, BookingId } from '../../domain/model/booking/booking';
-
-export interface BookingRepository {
-  nextId: () => BookingId;
-  findById: (id: BookingId) => Booking | undefined;
-  save: (booking: Booking) => void;
-}
