+++ src/booking/usecase/room/index.ts
@@ -1,4 +1,3 @@
 export * from './search';
 export * from './create';
-export * from './repository';
 export * from './edit';
